+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+v1.1.0-dev
+==========
+
+ * Added stats module
+ * [stats] added mean geohash
+ * [stats] added northernmost
+ * [stats] added southernmost
+ * [stats] added westernmost
+ * [stats] added easternmost
+
 v1.0.0
 ======